Fixed the size computation of both FileInputStream and FileOutputStream on Windows.
SEEK_END is not required to be implemented in a meaningful away
according to the C90 standard. On Windows, some compilers like GCC do
not implement SEEK_END and hence we need a different approach here. This
issue is potentially also present on other OS / compiler combinations
but the Posix way of doing this using the stat function requires the
path of the file rather than a file handle which means that that
approach is not quite compatible with how these classes are currently
implemented. I am not aware of any other good approach so the unsafe one
has to suffice for now. In the worst case, the only C portable method of
getting the size of a file is by reading the file until EOF and counting
the bytes.

Also I added a .gitattributes which should help avoid CRLF line endings.
For people contributing on Windows.